server/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.4.1
eventlet==0.36.1
pymongo==4.10.1
python-dotenv==1.0.1
requests==2.32.3
geopy==2.4.1
pytest==8.3.3
